Update package-lock.json
The previous package-lock.json was probably generated with npm v5, which pins
the exact version in the "requires" of a dependency. As of npm v6, only the
"version" is pinned to the exact version, and the "requires" follow what
package.json says (e.g. ^7.0.0 instead of always 7.0.0).

This commit also adds sequelize-cli and its dependencies to package-lock.json.